Extract brush compile artifact helpers
